uint32t

+++ com_prop_board.h
@@ -73,19 +73,19 @@ typedef struct __attribute__((__packed__)) {
     uint8_t ETH_main_valves_homing_done = false;
     uint8_t N2O_main_valves_homing_done = false;
     
-    uint16_t pid_eth_mass_flow_measure = 0.0;
-    uint16_t pid_eth_mass_flow_error = 0.0;
-    uint16_t pid_eth_mass_flow_error_integration = 0.0;
-    uint16_t pid_eth_mass_flow_error_derivative = 0.0;
-    uint16_t pid_eth_position_valve_cmd_feed_forward = 0.0;
-    uint16_t pid_eth_position_valve_cmd_pid = 0.0;
+    uint32_t pid_eth_mass_flow_measure = 0.0;
+    uint32_t pid_eth_mass_flow_error = 0.0;
+    uint32_t pid_eth_mass_flow_error_integration = 0.0;
+    uint32_t pid_eth_mass_flow_error_derivative = 0.0;
+    uint32_t pid_eth_position_valve_cmd_feed_forward = 0.0;
+    uint32_t pid_eth_position_valve_cmd_pid = 0.0;
 
-    uint16_t pid_n2o_mass_flow_measure = 0.0;
-    uint16_t pid_n2o_mass_flow_error = 0.0;
-    uint16_t pid_n2o_mass_flow_error_integration = 0.0;
-    uint16_t pid_n2o_mass_flow_error_derivative = 0.0;
-    uint16_t pid_n2o_position_valve_cmd_feed_forward = 0.0;
-    uint16_t pid_n2o_position_valve_cmd_pid = 0.0;
+    uint32_t pid_n2o_mass_flow_measure = 0.0;
+    uint32_t pid_n2o_mass_flow_error = 0.0;
+    uint32_t pid_n2o_mass_flow_error_integration = 0.0;
+    uint32_t pid_n2o_mass_flow_error_derivative = 0.0;
+    uint32_t pid_n2o_position_valve_cmd_feed_forward = 0.0;
+    uint32_t pid_n2o_position_valve_cmd_pid = 0.0;
 
     
     uint16_t temp_N2O = 0;          // [-70.0;40.0]
